#212 - Moved MapContext to context directory
