+++ apps/frontend/__tests__/MapMetaData.spec.tsx
@@ -4,7 +4,7 @@ import MapMetaData from '../src/app/components/MapMetaData';
 import { act, fireEvent, render } from '@testing-library/react';
 import { fetchItems, insertDatalayerView } from '../src/app/services/api';
 import { changeLayer } from '../src/app/components/MapView';
-import { useMapLayerContext } from '../src/app/components/MapContext';
+import { useMapLayerContext } from '../src/context/MapContext';
 
 jest.mock('sweetalert2-react-content', () => {
   return jest.fn().mockImplementation(() => ({
