+++ apps/frontend/src/app/components/SettingsPanel.tsx
@@ -18,7 +18,7 @@ import { toast, ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 import Swal from 'sweetalert2';
 import withReactContent from 'sweetalert2-react-content';
-import { useMapLayerContext } from './MapContext';
+import { useMapLayerContext } from '../../context/MapContext';
 import { getConfig, saveConfig } from '../services/configApi';
 
 const MySwal = withReactContent(Swal);
